refactor(domain): split path utils and table formatter into dedicated modules
Move pure path-manipulation helpers and the worktree-list table formatter
out of `src/domain/worktree.rs` so the domain layer no longer depends on
the display, color, or chrono concerns it was importing.

- Create `src/path_utils.rs` (top-level) housing `canonicalize_allow_missing`,
  `display_path`, `normalize_absolute_path` (all `pub`), and the private
  `normalize_path_lexically` helper. `canonicalize_allow_missing` was previously
  in `src/commands/common.rs`; the other three lived in `src/domain/worktree.rs`.
  Co-locating the two functions made a duplicate component-walk loop visible:
  `canonicalize_allow_missing` now reuses `normalize_path_lexically` instead of
  inlining the same `match Component {...}` block (subtractive, behavior-preserving).
- Create `src/commands/list_display.rs` housing `format_worktree_table` plus its
  private `WorktreeDisplay` struct. The sole consumer is `commands/list.rs`.
- Update all callers (`commands/{add,cd,common,create,init,list,open,rm}.rs`,
  `domain/worktree.rs`, `integrations/fzf.rs`) to import from the new locations.
  Combined imports that mixed moved and retained symbols are split into two
  `use` statements so the dependency direction is explicit.
- Drop `chrono`, `chrono_humanize`, `crate::color`, and `crate::commands::*`
  imports from `src/domain/`.
- 28 unit tests relocated alongside their owners (20 to path_utils, 8 to
  list_display); test count and behavior unchanged at 619 passing.
