+++ src/domain/worktree.rs
@@ -2,11 +2,7 @@
 //!
 //! This module contains data structures and parsing logic for git worktrees.
 
-use chrono::{DateTime, Utc};
-use chrono_humanize::{Accuracy, HumanTime, Tense};
-
-use crate::color;
-use crate::commands::common::canonicalize_allow_missing;
+use crate::path_utils::canonicalize_allow_missing;
 
 /// Worktree entry for enhanced display
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -15,16 +11,6 @@ pub struct WorktreeEntry {
     pub branch: Option<String>,
     pub hash: Option<String>,
     pub is_active: bool,
-}
-
-/// Worktree display information including commit time
-struct WorktreeDisplay {
-    path: Option<String>,
-    hash: String,
-    rel_path: Option<String>,
-    branch: String,
-    timestamp: String,
-    is_active: bool,
 }
 
 /// Check if a worktree path matches the active path
@@ -169,84 +155,6 @@ impl WorktreeList {
     pub fn current(&self) -> Option<&WorktreeEntry> {
         self.entries.iter().find(|e| e.is_active)
     }
-}
-
-/// Lexically normalize a path by resolving `.` and `..` components
-///
-/// Does NOT resolve symlinks or touch the filesystem
-fn normalize_path_lexically(path: &std::path::Path) -> std::path::PathBuf {
-    use std::path::Component;
-
-    let mut normalized = std::path::PathBuf::new();
-    for component in path.components() {
-        match component {
-            Component::CurDir => {
-                // Skip "." components
-            }
-            Component::ParentDir => {
-                // Pop the last component for ".."
-                normalized.pop();
-            }
-            _ => {
-                // Normal component (RootDir, Prefix, or Normal)
-                normalized.push(component);
-            }
-        }
-    }
-    normalized
-}
-
-/// Convert absolute path to home-relative display format
-///
-/// Returns "~/path" if under home directory, otherwise absolute path
-/// Normalizes paths lexically by resolving `.` and `..` components
-#[must_use]
-pub fn display_path(path: &std::path::Path) -> String {
-    // Normalize the path lexically (without resolving symlinks)
-    let normalized = normalize_path_lexically(path);
-
-    if let Some(home) = dirs::home_dir() {
-        if let Ok(rel) = normalized.strip_prefix(&home) {
-            let rel_str = rel.display().to_string();
-            if rel_str.is_empty() {
-                return "~".to_string();
-            }
-            return format!("~/{rel_str}");
-        }
-    }
-    normalized.display().to_string()
-}
-
-/// Normalize path to absolute form without tilde conversion
-///
-/// Resolves `.` and `..` components lexically.
-/// Use this for programmatic output (stdout) where absolute paths are needed.
-/// For human-friendly display, use `display_path()` instead.
-///
-/// If the input path is relative, it will be converted to an absolute path
-/// first using `canonicalize_allow_missing`, which works even if the path
-/// doesn't exist on the filesystem.
-///
-/// # Examples
-///
-/// ```
-/// use std::path::Path;
-/// use ofsht::domain::worktree::normalize_absolute_path;
-///
-/// let path = Path::new("/Users/test/ofsht/../ofsht-worktrees/feature");
-/// assert_eq!(
-///     normalize_absolute_path(path),
-///     "/Users/test/ofsht-worktrees/feature"
-/// );
-/// ```
-#[must_use]
-pub fn normalize_absolute_path(path: &std::path::Path) -> String {
-    let abs_path = if path.is_absolute() {
-        path.to_path_buf()
-    } else {
-        canonicalize_allow_missing(path)
-    };
-    normalize_path_lexically(&abs_path).display().to_string()
 }
 
 /// Calculate the depth from {branch} placeholder to the worktree root
@@ -405,343 +313,10 @@ pub fn calculate_relative_path(
         .map(|p| p.display().to_string())
 }
 
-/// Format worktree entries as a table with aligned columns
-///
-/// Returns formatted lines ready for display
-/// If `show_path` is false and `config` is None: hash • branch • time
-/// If `show_path` is false and `config` is Some: hash • `rel_path` • branch • time
-/// If `show_path` is true: path • hash • `rel_path` • branch • time
-///
-/// # Panics
-/// Panics if entries and `commit_times` have different lengths
-#[must_use]
-#[allow(clippy::too_many_lines)]
-pub fn format_worktree_table(
-    entries: &[WorktreeEntry],
-    commit_times: &[Option<DateTime<Utc>>],
-    show_path: bool,
-    color_mode: color::ColorMode,
-    config: Option<&crate::config::Config>,
-) -> Vec<String> {
-    assert_eq!(
-        entries.len(),
-        commit_times.len(),
-        "Entries and commit times must have same length"
-    );
-
-    let now = Utc::now();
-    let mut displays: Vec<WorktreeDisplay> = Vec::new();
-
-    // Calculate worktree root if config is provided
-    // Collect all non-main worktree paths (skip index 0 which is main worktree)
-    let worktree_root = config.and_then(|_cfg| {
-        let non_main_paths: Vec<std::path::PathBuf> = entries
-            .iter()
-            .skip(1)
-            .map(|entry| std::path::PathBuf::from(&entry.path))
-            .collect();
-
-        calculate_worktree_root_from_paths(&non_main_paths)
-    });
-
-    // Build display data
-    for (index, (entry, commit_time)) in entries.iter().zip(commit_times.iter()).enumerate() {
-        let path = if show_path {
-            Some(display_path(&std::path::PathBuf::from(&entry.path)))
-        } else {
-            None
-        };
-        let hash = entry
-            .hash
-            .clone()
-            .unwrap_or_else(|| "(unknown)".to_string());
-
-        // Calculate relative path for non-main worktrees
-        let rel_path = if index != 0 {
-            worktree_root.as_ref().and_then(|root| {
-                calculate_relative_path(&std::path::PathBuf::from(&entry.path), root)
-            })
-        } else {
-            None
-        };
-
-        // Main worktree (index 0) is always displayed as "@"
-        let branch = if index == 0 {
-            "[@]".to_string()
-        } else {
-            entry
-                .branch
-                .as_ref()
-                .map_or_else(|| "[detached]".to_string(), |b| format!("[{b}]"))
-        };
-        let timestamp = commit_time.as_ref().map_or_else(
-            || "–".to_string(),
-            |dt| {
-                let duration = now.signed_duration_since(*dt);
-                HumanTime::from(duration).to_text_en(Accuracy::Rough, Tense::Past)
-            },
-        );
-
-        displays.push(WorktreeDisplay {
-            path,
-            hash,
-            rel_path,
-            branch,
-            timestamp,
-            is_active: entry.is_active,
-        });
-    }
-
-    // Calculate column widths
-    let max_path_width = if show_path {
-        displays
-            .iter()
-            .filter_map(|d| d.path.as_ref().map(std::string::String::len))
-            .max()
-            .unwrap_or(0)
-    } else {
-        0
-    };
-    let max_hash_width = displays.iter().map(|d| d.hash.len()).max().unwrap_or(0);
-    let max_rel_path_width = displays
-        .iter()
-        .filter_map(|d| d.rel_path.as_ref().map(std::string::String::len))
-        .max()
-        .unwrap_or(0);
-    let max_branch_width = displays.iter().map(|d| d.branch.len()).max().unwrap_or(0);
-
-    // Format lines with padding and colors
-    displays
-        .iter()
-        .enumerate()
-        .map(|(index, d)| {
-            // Create active marker
-            let marker = if d.is_active {
-                color_mode.colorize_active_marker("*")
-            } else {
-                " ".to_string()
-            };
-
-            // Apply colors to each component
-            let colored_branch = if index == 0 {
-                // Main worktree [@] in green
-                color_mode.colorize_main_worktree(&d.branch)
-            } else if d.branch == "[detached]" {
-                // Detached HEAD in yellow
-                color_mode.colorize_detached(&d.branch)
-            } else {
-                // Regular branch in cyan
-                color_mode.colorize_branch(&d.branch)
-            };
-            let colored_timestamp = color_mode.colorize_secondary(&d.timestamp);
-
-            // Manual padding (format! doesn't work correctly with ANSI codes)
-            let hash_padding = " ".repeat(max_hash_width.saturating_sub(d.hash.len()));
-            let branch_padding = " ".repeat(max_branch_width.saturating_sub(d.branch.len()));
-
-            if show_path {
-                let colored_path = d.path.as_ref().unwrap();
-                let path_padding =
-                    " ".repeat(max_path_width.saturating_sub(colored_path.len()));
-
-                // Format relative path with padding
-                let rel_path_str = d.rel_path.as_deref().unwrap_or("");
-                let rel_path_padding =
-                    " ".repeat(max_rel_path_width.saturating_sub(rel_path_str.len()));
-
-                format!("{marker} {colored_path}{path_padding}  {}{hash_padding}  {rel_path_str}{rel_path_padding}  {colored_branch}{branch_padding}  {colored_timestamp}", d.hash)
-            } else if max_rel_path_width > 0 {
-                // Show relative path column when config is provided
-                let rel_path_str = d.rel_path.as_deref().unwrap_or("");
-                let rel_path_padding =
-                    " ".repeat(max_rel_path_width.saturating_sub(rel_path_str.len()));
-
-                format!("{marker} {}{hash_padding}  {rel_path_str}{rel_path_padding}  {colored_branch}{branch_padding}  {colored_timestamp}", d.hash)
-            } else {
-                // Original format without relative path
-                format!("{marker} {}{hash_padding}  {colored_branch}{branch_padding}  {colored_timestamp}", d.hash)
-            }
-        })
-        .collect()
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use std::path::PathBuf;
-
-    #[test]
-    fn test_normalize_path_lexically_removes_parent_dirs() {
-        // Test that .. components are resolved lexically
-        let path = PathBuf::from("/Users/test/ofsht/../ofsht-worktrees/feature");
-        let result = normalize_path_lexically(&path);
-        assert_eq!(result, PathBuf::from("/Users/test/ofsht-worktrees/feature"));
-    }
-
-    #[test]
-    fn test_normalize_path_lexically_removes_current_dirs() {
-        // Test that . components are skipped
-        let path = PathBuf::from("/Users/./test/./feature");
-        let result = normalize_path_lexically(&path);
-        assert_eq!(result, PathBuf::from("/Users/test/feature"));
-    }
-
-    #[test]
-    fn test_normalize_path_lexically_preserves_symlinks() {
-        // Test that symlinks are NOT resolved (lexical only)
-        // This test just verifies the function doesn't touch the filesystem
-        let path = PathBuf::from("/path/to/symlink/../target");
-        let result = normalize_path_lexically(&path);
-        assert_eq!(result, PathBuf::from("/path/to/target"));
-    }
-
-    #[test]
-    #[allow(clippy::literal_string_with_formatting_args)]
-    fn test_display_path_normalizes_parent_dirs() {
-        // Test that paths with .. are normalized
-        use std::path::MAIN_SEPARATOR;
-        let path = PathBuf::from(format!(
-            "{MAIN_SEPARATOR}Users{MAIN_SEPARATOR}test{MAIN_SEPARATOR}ofsht{MAIN_SEPARATOR}..{MAIN_SEPARATOR}ofsht-worktrees{MAIN_SEPARATOR}feature"
-        ));
-        let result = display_path(&path);
-        // Should not contain ../
-        assert!(!result.contains(".."));
-        // Should contain normalized path components
-        assert!(result.contains(&format!("ofsht-worktrees{MAIN_SEPARATOR}feature")));
-    }
-
-    #[test]
-    fn test_display_path_normalizes_home_relative_with_parent_dirs() {
-        // Test that home-relative paths with .. are normalized
-        if let Some(home) = dirs::home_dir() {
-            let path = home
-                .join("projects")
-                .join("ofsht")
-                .join("..")
-                .join("ofsht-worktrees")
-                .join("feature");
-            let result = display_path(&path);
-            // Should not contain ../
-            assert!(!result.contains(".."));
-            // Should contain normalized path
-            let sep = std::path::MAIN_SEPARATOR;
-            assert!(
-                result.contains(&format!("ofsht-worktrees{sep}feature"))
-                    || result.contains("ofsht-worktrees/feature")
-            ); // Unix-style in tilde paths
-        }
-    }
-
-    #[test]
-    #[allow(clippy::literal_string_with_formatting_args)]
-    fn test_display_path_outside_home() {
-        // Test paths outside home directory
-        use std::path::MAIN_SEPARATOR;
-        let path = if cfg!(windows) {
-            PathBuf::from(format!("C:{MAIN_SEPARATOR}temp{MAIN_SEPARATOR}worktree"))
-        } else {
-            PathBuf::from(format!("{MAIN_SEPARATOR}tmp{MAIN_SEPARATOR}worktree"))
-        };
-        let result = display_path(&path);
-        assert!(!result.starts_with('~'));
-    }
-
-    #[test]
-    #[allow(clippy::literal_string_with_formatting_args)]
-    fn test_normalize_absolute_path_resolves_parent_dirs() {
-        // Test that .. components are resolved
-        use std::path::MAIN_SEPARATOR;
-        let path = if cfg!(windows) {
-            PathBuf::from(format!(
-                "C:{MAIN_SEPARATOR}Users{MAIN_SEPARATOR}test{MAIN_SEPARATOR}ofsht{MAIN_SEPARATOR}..{MAIN_SEPARATOR}ofsht-worktrees{MAIN_SEPARATOR}feature"
-            ))
-        } else {
-            PathBuf::from("/Users/test/ofsht/../ofsht-worktrees/feature")
-        };
-        let result = normalize_absolute_path(&path);
-        let expected = if cfg!(windows) {
-            format!("C:{MAIN_SEPARATOR}Users{MAIN_SEPARATOR}test{MAIN_SEPARATOR}ofsht-worktrees{MAIN_SEPARATOR}feature")
-        } else {
-            "/Users/test/ofsht-worktrees/feature".to_string()
-        };
-        assert_eq!(result, expected);
-        assert!(!result.contains(".."));
-    }
-
-    #[test]
-    #[allow(clippy::literal_string_with_formatting_args)]
-    fn test_normalize_absolute_path_removes_current_dirs() {
-        // Test that . components are removed
-        use std::path::MAIN_SEPARATOR;
-        let path = if cfg!(windows) {
-            PathBuf::from(format!(
-                "C:{MAIN_SEPARATOR}Users{MAIN_SEPARATOR}.{MAIN_SEPARATOR}test{MAIN_SEPARATOR}.{MAIN_SEPARATOR}feature"
-            ))
-        } else {
-            PathBuf::from("/Users/./test/./feature")
-        };
-        let result = normalize_absolute_path(&path);
-        let expected = if cfg!(windows) {
-            format!("C:{MAIN_SEPARATOR}Users{MAIN_SEPARATOR}test{MAIN_SEPARATOR}feature")
-        } else {
-            "/Users/test/feature".to_string()
-        };
-        assert_eq!(result, expected);
-        // Check that . is removed (but be careful as it might appear in extensions)
-        assert!(!result.split(MAIN_SEPARATOR).any(|x| x == "."));
-    }
-
-    #[test]
-    #[allow(clippy::literal_string_with_formatting_args)]
-    fn test_normalize_absolute_path_outside_home() {
-        // Test paths outside home directory (no tilde conversion)
-        use std::path::MAIN_SEPARATOR;
-        let path = if cfg!(windows) {
-            PathBuf::from(format!("C:{MAIN_SEPARATOR}temp{MAIN_SEPARATOR}worktree"))
-        } else {
-            PathBuf::from("/tmp/worktree")
-        };
-        let result = normalize_absolute_path(&path);
-        assert!(!result.starts_with('~'));
-        // Verify path structure is preserved
-        assert!(result.contains("worktree"));
-    }
-
-    #[test]
-    #[allow(clippy::literal_string_with_formatting_args)]
-    fn test_normalize_absolute_path_consistency_with_display_path() {
-        // Test that normalization is consistent between the two functions
-        use std::path::MAIN_SEPARATOR;
-        let path = if cfg!(windows) {
-            PathBuf::from(format!(
-                "C:{MAIN_SEPARATOR}Users{MAIN_SEPARATOR}test{MAIN_SEPARATOR}ofsht{MAIN_SEPARATOR}..{MAIN_SEPARATOR}worktrees{MAIN_SEPARATOR}.{MAIN_SEPARATOR}feature{MAIN_SEPARATOR}.."
-            ))
-        } else {
-            PathBuf::from("/Users/test/ofsht/../worktrees/./feature/..")
-        };
-        let normalized_abs = normalize_absolute_path(&path);
-        let displayed = display_path(&path);
-
-        // Both should normalize the same way (only difference is tilde conversion)
-        assert!(!normalized_abs.contains(".."));
-        assert!(!normalized_abs.split(MAIN_SEPARATOR).any(|x| x == "."));
-        assert!(!displayed.contains(".."));
-
-        // Absolute result should not have tilde
-        assert!(!normalized_abs.starts_with('~'));
-    }
-
-    #[test]
-    fn test_normalize_absolute_path_handles_relative_paths() {
-        // Test that relative paths are safely converted to absolute paths
-        let relative = PathBuf::from("worktrees/feature");
-        let result = normalize_absolute_path(&relative);
-
-        // Should be converted to absolute path (works on both Unix and Windows)
-        assert!(PathBuf::from(&result).is_absolute());
-        assert!(result.contains("worktrees"));
-        assert!(result.contains("feature"));
-    }
 
     // --- Tests for relative path calculation utilities ---
 
@@ -825,329 +400,6 @@ mod tests {
             calculate_relative_path(&worktree, &root),
             Some("team/alice/fix".to_string())
         );
-    }
-
-    // --- Tests moved from src/main.rs (mechanical relocation, behavior preserved) ---
-
-    #[test]
-    fn test_display_path_under_home() {
-        // Test path under home directory
-        if let Some(home) = dirs::home_dir() {
-            let test_path = home.join("test/path");
-            let result = display_path(&test_path);
-            assert!(result.starts_with("~/"));
-            assert_eq!(result, "~/test/path");
-        }
-    }
-
-    #[test]
-    fn test_display_path_outside_home_unix_exact() {
-        // Test path outside home directory (Unix-only exact-match assertion).
-        // Renamed from `test_display_path_outside_home` because that name is already
-        // used above for a platform-aware (cfg!(windows)) version that only checks
-        // the leading '~' is absent. This keeps both assertions.
-        let test_path = std::path::PathBuf::from("/tmp/test/path");
-        let result = display_path(&test_path);
-        assert_eq!(result, "/tmp/test/path");
-    }
-
-    #[test]
-    fn test_display_path_home_itself() {
-        // Test home directory itself
-        if let Some(home) = dirs::home_dir() {
-            let result = display_path(&home);
-            assert_eq!(result, "~");
-        }
-    }
-
-    #[test]
-    fn test_format_worktree_table_default_no_path() {
-        let entries = vec![WorktreeEntry {
-            path: "/path/to/main".to_string(),
-            branch: Some("main".to_string()),
-            hash: Some("a1b2c3d4".to_string()),
-            is_active: false,
-        }];
-        let commit_times = vec![Some(
-            DateTime::from_timestamp(Utc::now().timestamp() - 3600, 0).unwrap(),
-        )];
-
-        let result = format_worktree_table(
-            &entries,
-            &commit_times,
-            false,
-            crate::color::ColorMode::Never,
-            None,
-        );
-        assert_eq!(result.len(), 1);
-        // Line should contain hash, branch, and timestamp (no path)
-        assert!(!result[0].contains("/path/to/main"));
-        // Main worktree (first entry) should be displayed as [@]
-        assert!(result[0].contains("[@]"));
-        assert!(result[0].contains("a1b2c3d4"));
-    }
-
-    #[test]
-    fn test_format_worktree_table_with_path() {
-        let entries = vec![WorktreeEntry {
-            path: "/path/to/main".to_string(),
-            branch: Some("main".to_string()),
-            hash: Some("a1b2c3d4".to_string()),
-            is_active: false,
-        }];
-        let commit_times = vec![Some(
-            DateTime::from_timestamp(Utc::now().timestamp() - 3600, 0).unwrap(),
-        )];
-
-        let result = format_worktree_table(
-            &entries,
-            &commit_times,
-            true,
-            crate::color::ColorMode::Never,
-            None,
-        );
-        assert_eq!(result.len(), 1);
-        // Line should contain path, hash, branch, and timestamp
-        assert!(result[0].contains("/path/to/main"));
-        // Main worktree (first entry) should be displayed as [@]
-        assert!(result[0].contains("[@]"));
-        assert!(result[0].contains("a1b2c3d4"));
-    }
-
-    #[test]
-    fn test_format_worktree_table_multiple_entries() {
-        let entries = vec![
-            WorktreeEntry {
-                path: "/path/to/main".to_string(),
-                branch: Some("main".to_string()),
-                hash: Some("a1b2c3d4".to_string()),
-                is_active: false,
-            },
-            WorktreeEntry {
-                path: "/path/to/feature-branch".to_string(),
-                branch: Some("feature".to_string()),
-                hash: Some("e5f6g7h8".to_string()),
-                is_active: false,
-            },
-        ];
-        let commit_times = vec![
-            Some(DateTime::from_timestamp(Utc::now().timestamp() - 3600, 0).unwrap()),
-            None,
-        ];
-
-        let result = format_worktree_table(
-            &entries,
-            &commit_times,
-            false,
-            crate::color::ColorMode::Never,
-            None,
-        );
-        assert_eq!(result.len(), 2);
-        // Both lines should have same structure (no paths)
-        assert!(!result[0].contains("/path/to/main"));
-        // Main worktree (first entry) should be displayed as [@]
-        assert!(result[0].contains("[@]"));
-        assert!(result[0].contains("a1b2c3d4"));
-        assert!(!result[1].contains("/path/to/feature-branch"));
-        assert!(result[1].contains("[feature]"));
-        assert!(result[1].contains("e5f6g7h8"));
-        assert!(result[1].contains("–")); // No commit time
-    }
-
-    #[test]
-    fn test_format_worktree_table_column_alignment() {
-        let entries = vec![
-            WorktreeEntry {
-                path: "/short".to_string(),
-                branch: Some("a".to_string()),
-                hash: Some("12345678".to_string()),
-                is_active: false,
-            },
-            WorktreeEntry {
-                path: "/very/long/path/to/worktree".to_string(),
-                branch: Some("feature-branch".to_string()),
-                hash: Some("abcdefgh".to_string()),
-                is_active: false,
-            },
-        ];
-        let commit_times = vec![None, None];
-
-        let result = format_worktree_table(
-            &entries,
-            &commit_times,
-            true,
-            crate::color::ColorMode::Never,
-            None,
-        );
-        assert_eq!(result.len(), 2);
-
-        // Verify that both lines contain the expected content
-        assert!(result[0].contains("/short"));
-        // Main worktree (first entry) should be displayed as [@]
-        assert!(result[0].contains("[@]"));
-        assert!(result[0].contains("12345678"));
-        assert!(result[1].contains("/very/long/path/to/worktree"));
-        assert!(result[1].contains("[feature-branch]"));
-        assert!(result[1].contains("abcdefgh"));
-
-        // Both lines should end with the same timestamp placeholder
-        assert!(result[0].ends_with("–"));
-        assert!(result[1].ends_with("–"));
-    }
-
-    #[test]
-    fn test_format_worktree_table_detached_head() {
-        let entries = vec![WorktreeEntry {
-            path: "/path/to/detached".to_string(),
-            branch: None,
-            hash: Some("deadbeef".to_string()),
-            is_active: false,
-        }];
-        let commit_times = vec![None];
-
-        let result = format_worktree_table(
-            &entries,
-            &commit_times,
-            false,
-            crate::color::ColorMode::Never,
-            None,
-        );
-        assert_eq!(result.len(), 1);
-        // Main worktree (first entry) is always [@], even if detached
-        assert!(result[0].contains("[@]"));
-        assert!(result[0].contains("deadbeef"));
-        assert!(result[0].contains("–"));
-    }
-
-    #[test]
-    fn test_format_worktree_table_active_marker() {
-        let entries = vec![
-            WorktreeEntry {
-                path: "/path/to/main".to_string(),
-                branch: Some("main".to_string()),
-                hash: Some("a1b2c3d4".to_string()),
-                is_active: false,
-            },
-            WorktreeEntry {
-                path: "/path/to/feature".to_string(),
-                branch: Some("feature".to_string()),
-                hash: Some("e5f6g7h8".to_string()),
-                is_active: true,
-            },
-        ];
-        let commit_times = vec![None, None];
-
-        let result = format_worktree_table(
-            &entries,
-            &commit_times,
-            false,
-            crate::color::ColorMode::Never,
-            None,
-        );
-        assert_eq!(result.len(), 2);
-        // First entry (inactive) should have space prefix
-        assert!(result[0].starts_with("  "));
-        // Second entry (active) should have * prefix
-        assert!(result[1].starts_with("* "));
-        assert!(result[1].contains("[feature]"));
-    }
-
-    #[test]
-    fn test_format_worktree_table_active_marker_with_path() {
-        let entries = vec![
-            WorktreeEntry {
-                path: "/path/to/main".to_string(),
-                branch: Some("main".to_string()),
-                hash: Some("a1b2c3d4".to_string()),
-                is_active: false,
-            },
-            WorktreeEntry {
-                path: "/path/to/feature".to_string(),
-                branch: Some("feature".to_string()),
-                hash: Some("e5f6g7h8".to_string()),
-                is_active: true,
-            },
-        ];
-        let commit_times = vec![None, None];
-
-        let result = format_worktree_table(
-            &entries,
-            &commit_times,
-            true,
-            crate::color::ColorMode::Never,
-            None,
-        );
-        assert_eq!(result.len(), 2);
-        // Both entries should have marker prefix (space or *)
-        assert!(result[0].starts_with("  "));
-        assert!(result[1].starts_with("* "));
-    }
-
-    #[test]
-    fn test_format_worktree_table_with_relative_paths() {
-        // Test that relative paths are displayed when config is provided
-        use crate::config::{Config, Hooks, IntegrationsConfig, WorktreeConfig};
-
-        let entries = vec![
-            WorktreeEntry {
-                path: "/Users/test/repo-worktrees/main".to_string(),
-                branch: Some("main".to_string()),
-                hash: Some("a1b2c3d4".to_string()),
-                is_active: false,
-            },
-            WorktreeEntry {
-                path: "/Users/test/repo-worktrees/feature".to_string(),
-                branch: Some("feature".to_string()),
-                hash: Some("e5f6g7h8".to_string()),
-                is_active: false,
-            },
-            WorktreeEntry {
-                path: "/Users/test/repo-worktrees/docs/tweak".to_string(),
-                branch: Some("docs/tweak".to_string()),
-                hash: Some("i9j0k1l2".to_string()),
-                is_active: true,
-            },
-        ];
-        let commit_times = vec![None, None, None];
-
-        let config = Config {
-            worktree: WorktreeConfig {
-                dir: "../{repo}-worktrees/{branch}".to_string(),
-            },
-            hooks: Hooks::default(),
-            integrations: IntegrationsConfig::default(),
-        };
-
-        let result = format_worktree_table(
-            &entries,
-            &commit_times,
-            false,
-            crate::color::ColorMode::Never,
-            Some(&config),
-        );
-
-        assert_eq!(result.len(), 3);
-
-        // Main worktree (index 0) should have blank relative path column
-        // Format: "  a1b2c3d4            [@]  unknown" (extra spaces where rel_path would be)
-        let main_line = &result[0];
-        assert!(main_line.contains("a1b2c3d4"));
-        assert!(main_line.contains("[@]"));
-
-        // Feature worktree should show "feature" as relative path
-        // Format: "  e5f6g7h8  feature  [feature]  unknown"
-        let feature_line = &result[1];
-        assert!(feature_line.contains("e5f6g7h8"));
-        assert!(feature_line.contains("feature"));
-        assert!(feature_line.contains("[feature]"));
-
-        // Nested worktree should show "docs/tweak" as relative path
-        // Format: "* i9j0k1l2  docs/tweak  [docs/tweak]  unknown"
-        let nested_line = &result[2];
-        assert!(nested_line.contains("i9j0k1l2"));
-        assert!(nested_line.contains("docs/tweak"));
-        assert!(nested_line.contains("[docs/tweak]"));
-        assert!(nested_line.starts_with("* ")); // Active marker
     }
 
     // -----------------------------------------------------------------
